+++ src/common/common.ts
@@ -5,13 +5,13 @@ import { JsonCodec } from '../codecs/json/json.js';
 import { deleteContent, getContent, putContent } from '../content/api.js';
 import type { PutRequestPayload } from '../content/procedures.js';
 import { ECDSA } from '../ecdsa/wrap.js';
-import { EncryptWrapModule } from '../encrypt/wrap.js';
+import { CryptWrapModule } from '../crypt/index.js';
+import { WithWebCrypto } from '../crypt/web-crypto.js';
 import { parseAsFile } from '../file/parse.js';
 import { SHA_256, sha256 } from '../hashing/algorithms/sha256.js';
 import { prefix as identityPrefix, scheme as identityScheme } from '../identity/identity.js';
-import { IMMUTABLE_PREFIX } from '../immutable/repository.js';
-import { Immutable } from '../immutable/scheme.js';
-import type { InstanceConfig } from '../instance/instance.js';
+import { Immutable, IMMUTABLE_PREFIX } from '../immutable/index.js';
+import { createInstance, type Instance } from '../instance/instance.js';
 import { Binary, JSON as JsonMediaType, Wrap } from '../media-types/media-types.js';
 import { MUTABLE_PREFIX } from '../mutable/mutable.js';
 import { WrapCodec } from '../wraps/codec.js';
@@ -34,20 +34,20 @@ import { WrapCodec } from '../wraps/codec.js';
  *
  *     const instance = createInstance(Common);
  */
-export const Common = {
+export const Common: Instance = createInstance(WithWebCrypto, {
   codecs: {
     [JsonMediaType]: JsonCodec,
     [Binary]: BinaryCodec,
     [Wrap]: WrapCodec,
+  },
+  hashAlgs: {
+    [SHA_256]: sha256,
   },
   procedures: {
     'content:delete': deleteContent,
     'content:get': getContent,
     'content:put': (req, instance) =>
       putContent((req as PutRequestPayload).cid, (req as PutRequestPayload).content, { instance }),
-  },
-  hashAlgs: {
-    [SHA_256]: sha256,
   },
   schemes: {
     [IMMUTABLE_PREFIX]: Immutable,
@@ -56,6 +56,6 @@ export const Common = {
   },
   wraps: {
     ECDSA,
-    encrypt: EncryptWrapModule,
+    crypt: CryptWrapModule,
   },
-} as const satisfies InstanceConfig;
+});