+++ src/crypt/index.ts
@@ -1,0 +1,11 @@
+/**
+ * @module Encrypt
+ * @experimental
+ */
+
+export * from './defaults.js';
+export * from './encrypt.js';
+export * from './kdf.js';
+export * from './options.js';
+export * from './types.js';
+export * from './wrap.js';