+++ src/encrypt/index.ts
@@ -1,4 +1,0 @@
-/** @module Encrypt */
-
-export * from './encrypt.js';
-export * from './wrap.js';