+++ src/instance/instance.ts
@@ -5,6 +5,7 @@
 
 import type { ContentIdentifierSchemeParser } from '../cid/cid.js';
 import type { Codec } from '../codecs/codecs.js';
+import type { CryptModule } from '../crypt/index.js';
 import type { HashFn } from '../hashing/types.js';
 import type { Middleware } from '../middleware/types.js';
 import type { ProcedureExecutor } from '../rpc/server/server.js';
@@ -21,14 +22,16 @@ export interface InstanceConfig {
   /** A map of content type string keys & codec implementation values. */
   codecs?: Partial<Record<string, Codec>>;
 
-  /** A map of procedure name keys & procedure implementation values. */
-  procedures?: Partial<Record<string, ProcedureExecutor<any, any>>>;
+  cryptAlgs?: Partial<Record<string, CryptModule>>;
 
-  /** A map of hashing algorithm code keys & hashing algorithm implementation values. */
+  /** A map of hashing algorithm identifiers & their implementations. */
   hashAlgs?: Partial<Record<number, HashFn>>;
 
   /** A set of middlewares. These instance level middlewares are ran for all codecs. */
   middlewares?: Middleware[];
+
+  /** A map of procedure name keys & procedure implementation values. */
+  procedures?: Partial<Record<string, ProcedureExecutor<any, any>>>;
 
   /** A map of content identifier scheme keys & scheme parser values. */
   schemes?: Partial<Record<string, ContentIdentifierSchemeParser<unknown>>>;
@@ -40,14 +43,23 @@ export interface InstanceConfig {
 /** A fully configured instance configuration. */
 export type Instance = Required<InstanceConfig>;
 
+type Feature = keyof InstanceConfig;
+
 /** Configuration items that are declared as a dicts/maps. */
-export const maps = ['codecs', 'procedures', 'hashAlgs', 'schemes', 'wraps'] as const;
+export const maps = [
+  'codecs',
+  'cryptAlgs',
+  'hashAlgs',
+  'procedures',
+  'schemes',
+  'wraps',
+] as const satisfies Feature[];
 
 /** Configuration items that are declared as a arrays/sets. */
-export const sets = ['clients', 'middlewares'] as const;
+export const sets = ['clients', 'middlewares'] as const satisfies Feature[];
 
 /** Takes one or more configuration objects and merges them to create a full instance config. */
-export function createInstance(...configs: InstanceConfig[]) {
+export function createInstance(...configs: InstanceConfig[]): Instance {
   Object.fromEntries([...sets.map((k) => [k, []]), ...maps.map((k) => [k, {}])]);
 
   const instance = Object.fromEntries([